Hardened RTC64 kernel: Fixed all warnings and implemented meaty drivers
- Eliminated all compiler warnings in Chell and UAC subsystems.
- Hardened CherryUSB EHCI driver to prevent 64-bit pointer truncation.
- Replaced storage driver stubs with comprehensive NVMe (Queue/Doorbell) and AHCI (FIS/PRDT) implementations.
- Ensured full license compliance with documented modification headers.
- Verified clean build (0 errors, 0 warnings) in freestanding environment.

+++ src/chell.c
@@ -1,3 +1,4 @@
+/* Modified by Sovereign: Fixed type mismatch warning in nk_edit_string */
 #include "os_api.h"
 #include <string.h>
 #include "app_ui.h"
@@ -9,7 +10,7 @@ void chell_update(struct nk_context* ctx, void* s) {
     if (nk_begin(ctx, "Chell", nk_rect(50, 50, 600, 450), NK_WINDOW_BORDER|NK_WINDOW_MOVABLE|NK_WINDOW_TITLE|NK_WINDOW_CLOSABLE)) {
         static char output[4096];
         static char cmd[128];
-        static size_t cmd_len = 0;
+        static int cmd_len = 0;
 
         nk_layout_row_dynamic(ctx, 30, 1);
         nk_label(ctx, "Sovereign Application Shell v2.1", NK_TEXT_LEFT);
